Initial commit
This was started based on the fine work done at:

https://gitlab.com/asacolips-projects/foundry-mods/dungeonworld

I am starting it as a copy, not a fork, so I could remove assets and
content that I do not intend to reference, just to reduce the amound of
licensing, and identification of ownership of trademarks.

This approach loses some history in this version of the repository, but
I will keep a copy for my own referencing when needed.