Update license.txt, for multi-licenseness
The file module/paranoidworld.js is marked as going by GNU GPL v3,
because the original Dungeon World project had it labeled as such,
and I just carried that over as required.

To deal with this, I put both licenses (MIT for the majority of the
code) in license.txt, and noted which license applies where.

Additionally, credit the original authors. At least the ones who've put
their names in the files somewhere.